requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.13.0
openpyxl>=3.1.0
gspread>=5.12.0
google-auth>=2.23.0
requests>=2.28.0
google-cloud-storage>=2.10.0
